Split the hoist test module along its suite boundaries
The workflow-contract hoist tests had grown to 901 lines against the
repository's 400-line cap. Divide the module by concern, with no test
behaviour changed:

- `conftest.py` holds the shared `workspace` fixture, the `stage_pair`,
  `run_hoist`, and `assert_nothing_moved` helpers, the shared staging
  constants, and the `sys.path` seam to the production script.
- `hoist_binstall_archives_test.py` keeps the example-based naming,
  validation-failure, filesystem-safety, and CLI tests.
- `hoist_binstall_rollback_test.py` holds the move-rollback and
  failed-rollback transaction tests.
- `hoist_binstall_archives_generated_test.py` holds the Hypothesis
  generated-layout suite with its strategies and bounds unchanged.
- `release_workflow_hoist_test.py` holds the `release.yml` wiring
  contracts.
