+++ tests/workflow_contracts/hoist_binstall_rollback_test.py
@@ -1,0 +1,125 @@
+"""Rollback-transaction tests for the cargo-binstall hoist step.
+
+The move phase is all-or-nothing: a mid-phase failure must restore every
+completed move so the release job can rerun after the fault clears, and a
+failed restoration must surface both errors rather than losing the cause.
+These tests inject faults at the ``shutil.move`` seam to prove both
+properties. Shared fixtures and helpers live in ``conftest.py``.
+
+Run via ``make test-workflow-contracts``.
+"""
+
+from __future__ import annotations
+
+import shutil
+from pathlib import Path
+from unittest import mock
+
+import pytest
+from conftest import EXPECTED_NAMES, run_hoist, stage_pair
+
+# conftest.py inserts scripts/ onto sys.path before this module is imported.
+import hoist_binstall_archives as hoist_mod
+
+
+def test_hoist_rolls_back_completed_moves_when_a_move_fails(
+    workspace: dict[str, Path],
+) -> None:
+    """A mid-phase move failure restores every file, and a rerun succeeds."""
+    stage_pair(workspace["dist"], "netsuke-linux-amd64/s1", EXPECTED_NAMES[1])
+    stage_pair(workspace["dist"], "netsuke-macos-arm64/s2", EXPECTED_NAMES[0])
+    sources = {
+        EXPECTED_NAMES[0]: workspace["dist"] / "netsuke-macos-arm64/s2",
+        EXPECTED_NAMES[1]: workspace["dist"] / "netsuke-linux-amd64/s1",
+    }
+
+    real_move = shutil.move
+    calls = {"count": 0}
+
+    def failing_move(src, dst):  # noqa: ANN001, ANN202 - shutil.move signature
+        calls["count"] += 1
+        if calls["count"] == 3:
+            msg = "injected move failure"
+            raise OSError(msg)
+        return real_move(src, dst)
+
+    # The patch is scoped to the failing run so the retry below exercises the
+    # real `shutil.move`, proving the rollback left a rerunnable tree.
+    with (
+        mock.patch.object(hoist_mod.shutil, "move", failing_move),
+        pytest.raises(OSError, match="injected move failure"),
+    ):
+        run_hoist(workspace)
+
+    for name, nested in sources.items():
+        assert not (workspace["dist"] / name).exists(), (
+            f"{name} must not remain at the release root after rollback"
+        )
+        assert not (workspace["dist"] / f"{name}.sha256").exists(), (
+            f"{name}.sha256 must not remain at the release root after rollback"
+        )
+        restored = nested / name
+        assert restored.is_file(), f"{name} must return to its nested path"
+        assert restored.read_bytes() == f"archive:{name}".encode(), (
+            f"{name} must keep its original content through the rollback"
+        )
+        sidecar = nested / f"{name}.sha256"
+        assert sidecar.is_file(), f"{name}.sha256 must return to its nested path"
+        assert sidecar.read_text(encoding="utf-8") == f"checksum:{name}", (
+            f"{name}.sha256 must keep its original content through the rollback"
+        )
+
+    assert run_hoist(workspace) == 0, "a rerun after the fault clears must succeed"
+    for name in EXPECTED_NAMES:
+        assert (workspace["dist"] / name).is_file(), (
+            f"{name} must reach the release root on the retry"
+        )
+        assert (workspace["dist"] / name).read_bytes() == (
+            f"archive:{name}".encode()
+        ), f"{name} content must survive the retry"
+
+
+def test_hoist_surfaces_both_errors_when_the_rollback_also_fails(
+    workspace: dict[str, Path],
+) -> None:
+    """A failed rollback reports its own error alongside the original.
+
+    Losing the move failure to the rollback failure would leave an operator
+    with no way to tell why the release aborted, so both must survive.
+    """
+    stage_pair(workspace["dist"], "netsuke-linux-amd64/s1", EXPECTED_NAMES[1])
+    stage_pair(workspace["dist"], "netsuke-macos-arm64/s2", EXPECTED_NAMES[0])
+
+    real_move = shutil.move
+    calls = {"count": 0}
+
+    def failing_move(src, dst):  # noqa: ANN001, ANN202 - shutil.move signature
+        calls["count"] += 1
+        # Calls 1 and 2 relocate the first pair; call 3 aborts the forward
+        # phase; call 4 is the first restoration attempt and fails too.
+        if calls["count"] == 3:
+            msg = "injected move failure"
+            raise OSError(msg)
+        if calls["count"] > 3:
+            msg = "injected rollback failure"
+            raise OSError(msg)
+        return real_move(src, dst)
+
+    with (
+        mock.patch.object(hoist_mod.shutil, "move", failing_move),
+        pytest.raises(BaseExceptionGroup) as raised,
+    ):
+        run_hoist(workspace)
+
+    group = raised.value
+    assert "rollback could not restore every file" in str(group), (
+        f"the group must explain that the rollback failed; got {group}"
+    )
+    assert sorted(str(error) for error in group.exceptions) == [
+        "injected move failure",
+        "injected rollback failure",
+    ], f"both failures must survive in the group; got {group.exceptions}"
+    assert (workspace["dist"] / EXPECTED_NAMES[0]).is_file(), (
+        "an unrestorable pair must be left where it landed, not silently "
+        "discarded, so the operator can inspect the release root"
+    )
